schema version changed type

+++ plugin/conversions.go
@@ -13,7 +13,7 @@ import (
 // ProtoToProviderSchema takes a proto.Schema and converts it to a providers.Schema.
 func ProtoToProviderSchema(s *proto.Schema) providers.Schema {
 	return providers.Schema{
-		Version: int(s.Version),
+		Version: uint64(s.Version),
 		Block:   schemaBlock(s.Block),
 	}
 }
